+++ geopandas/io/file.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import os
 import urllib.request
 import warnings
+from http import HTTPStatus
 from io import IOBase
 from packaging.version import Version
 from pathlib import Path
@@ -10,10 +11,16 @@ from pathlib import Path
 # Adapted from pandas.io.common
 from urllib.parse import urlparse as parse_url
 from urllib.parse import uses_netloc, uses_params, uses_relative
+from urllib.request import Request
 
 import numpy as np
 import pandas as pd
-from pandas.api.types import is_datetime64_any_dtype, is_integer_dtype, is_object_dtype
+from pandas.api.types import (
+    is_datetime64_any_dtype,
+    is_integer_dtype,
+    is_object_dtype,
+    is_string_dtype,
+)
 
 import shapely
 from shapely.geometry import mapping
@@ -290,10 +297,21 @@ def _read_file(
         # pyogrio/fiona as is (to support downloading only part of the file)
         # otherwise still download manually because pyogrio/fiona don't support
         # all types of urls (https://github.com/geopandas/geopandas/issues/2908)
-        with urllib.request.urlopen(filename) as response:
-            if not response.headers.get("Accept-Ranges") == "bytes":
+        try:
+            with urllib.request.urlopen(
+                Request(filename, headers={"Range": "bytes=0-1"})
+            ) as response:
+                if (
+                    response.headers.get("Accept-Ranges") == "none"
+                    or response.status != HTTPStatus.PARTIAL_CONTENT
+                ):
+                    from_bytes = True
+        except ConnectionError:
+            from_bytes = True
+
+        if from_bytes:
+            with urllib.request.urlopen(filename) as response:
                 filename = response.read()
-                from_bytes = True
 
     if engine == "pyogrio":
         return _read_file_pyogrio(
@@ -361,7 +379,8 @@ def _read_file_fiona(
 
     with fiona_env():
         with reader(path_or_bytes, **kwargs) as features:
-            crs = features.crs_wkt
+            crs = features.crs_wkt  # returns "" if empty
+            crs = crs or None
             # attempt to get EPSG code
             try:
                 # fiona 1.9+
@@ -384,7 +403,11 @@ def _read_file_fiona(
                 assert len(bbox) == 4
             # handle loading the mask
             elif isinstance(mask, GeoDataFrame | GeoSeries):
-                mask = mapping(mask.to_crs(crs).union_all())
+                if crs is not None and mask.crs is not None:
+                    mask = mask.to_crs(crs)
+                else:
+                    _warn_missing_crs_of_dataframe_and_mask(crs, mask)
+                mask = mapping(mask.union_all())
             elif isinstance(mask, BaseGeometry):
                 mask = mapping(mask)
 
@@ -497,10 +520,15 @@ def _read_file_pyogrio(path_or_bytes, bbox=None, mask=None, rows=None, **kwargs)
         # NOTE: mask cannot be used at same time as bbox keyword
         if isinstance(mask, GeoDataFrame | GeoSeries):
             crs = pyogrio.read_info(path_or_bytes, layer=kwargs.get("layer")).get("crs")
+            if crs is not None and mask.crs is not None:
+                mask = mask.to_crs(crs)
+            else:
+                _warn_missing_crs_of_dataframe_and_mask(crs, mask)
+
             if isinstance(path_or_bytes, IOBase):
                 path_or_bytes.seek(0)
 
-            mask = shapely.unary_union(mask.to_crs(crs).geometry.values)
+            mask = shapely.unary_union(mask.geometry.values)
         elif isinstance(mask, BaseGeometry):
             mask = shapely.unary_union(mask)
         elif isinstance(mask, dict) or hasattr(mask, "__geo_interface__"):
@@ -547,6 +575,25 @@ def _read_file_pyogrio(path_or_bytes, bbox=None, mask=None, rows=None, **kwargs)
         kwargs["columns"] = kwargs.pop("include_fields")
 
     return pyogrio.read_dataframe(path_or_bytes, bbox=bbox, **kwargs)
+
+
+def _warn_missing_crs_of_dataframe_and_mask(source_dataset_crs, mask):
+    """
+    Warn if one, or both, of the source dataset or mask does not
+    have a crs.
+    """
+    if (source_dataset_crs is None) and (mask.crs is None):
+        msg = "There is no CRS defined in the source dataset nor mask. "
+    elif (source_dataset_crs is None) and (mask.crs is not None):
+        msg = "There is no CRS defined in the source dataset. "
+    else:  # crs not None and mask.crs is None
+        msg = "There is no CRS defined in the mask. "
+    msg += (
+        "This may lead to a misalignment of the mask and the "
+        "source dataset, leading to incorrect masking. Ensure "
+        "both inputs share the same CRS."
+    )
+    warnings.warn(msg, UserWarning, stacklevel=3)
 
 
 def _detect_driver(path):
@@ -763,7 +810,7 @@ def infer_schema(df):
     }
 
     def convert_type(column, in_type):
-        if is_object_dtype(in_type):
+        if is_object_dtype(in_type) or is_string_dtype(in_type):
             return "str"
         if is_datetime64_any_dtype(in_type):
             # numpy datetime type regardless of frequency
@@ -806,9 +853,10 @@ def _geometry_types(df):
     Determine the geometry types in the GeoDataFrame for the schema.
     """
     geom_types_2D = df[~df.geometry.has_z].geometry.geom_type.unique()
-    geom_types_2D = [gtype for gtype in geom_types_2D if gtype is not None]
+    geom_types_2D = list(geom_types_2D[pd.notna(geom_types_2D)])
     geom_types_3D = df[df.geometry.has_z].geometry.geom_type.unique()
-    geom_types_3D = ["3D " + gtype for gtype in geom_types_3D if gtype is not None]
+    geom_types_3D = list(geom_types_3D[pd.notna(geom_types_3D)])
+    geom_types_3D = ["3D " + gtype for gtype in geom_types_3D]
     geom_types = geom_types_3D + geom_types_2D
 
     if len(geom_types) == 0: